Move `core_media` into `media` crate
