+++ crates/capture/src/bindings.h
@@ -1,3 +1,2 @@
-#import <CoreMedia/CMSampleBuffer.h>
 #import <ScreenCaptureKit/SCStream.h>
 #import <dispatch/dispatch.h>
