+++ crates/media/src/bindings.h
@@ -1,2 +1,3 @@
+#import <CoreMedia/CMSampleBuffer.h>
 #import <CoreVideo/CVPixelFormatDescription.h>
 #import <CoreVideo/CVReturn.h>
